update load mount table

+++ voltdb/InsertMountEntries.java
@@ -5,7 +5,7 @@ public class InsertMountEntries extends VoltProcedure {
   public final SQLStmt sql =
       new SQLStmt("UPSERT INTO mount(namenode, path, readOnly) VALUES (?, ?, ?);");
 
-  public long run(final String[] namenodes, final String[] paths, final Integer[] readonlys)
+  public long run(final String[] namenodes, final String[] paths, final long[] readonlys)
       throws VoltAbortException {
     for (int i = 0; i < namenodes.length; ++i) {
       voltQueueSQL(sql, namenodes[i], paths[i], readonlys[i]);
